requirements.txt
opencv-python~=3.4.17.61
numpy~=1.22.0
djitellopy~=2.4.0
ttkbootstrap~=1.5.0
Pillow~=9.0.0
imutils~=0.5.4